+++ srcjs/amChart4.jsx
@@ -1990,9 +1990,23 @@ class AmLineChart extends React.PureComponent {
           );
         }
       });
-
-      
+      let zoomOutButton = buttonContainer.createChild(am4core.Button);
+      zoomOutButton.label.text = "-";
+      zoomOutButton.events.on("hit", function(ev) {
+        let diff = XAxis.maxZoomed - XAxis.minZoomed;
+        let delta = diff * zoomButtons.zoomFactor;
+        if(isDate){
+          XAxis.zoomToDates(
+            new Date(XAxis.minZoomed - delta), new Date(XAxis.maxZoomed + delta)
+          );
+        }else{
+          XAxis.zoomToValues(
+            XAxis.minZoomed - delta, XAxis.maxZoomed + delta
+          );
+        }
+      });
     }
+
 
     /* ~~~~\                         /~~~~ */
     yValues.forEach(function (value, index) {
@@ -2272,6 +2286,7 @@ class AmScatterChart extends React.PureComponent {
       draggable = this.props.draggable,
       pointsStyle = this.props.pointsStyle,
       cursor = this.props.cursor,
+      zoomButtons = this.props.zoomButtons,
       chartId = this.props.chartId,
       shinyId = this.props.shinyId;
 
@@ -2698,6 +2713,59 @@ class AmScatterChart extends React.PureComponent {
     });
 
 
+    /* ~~~~\  zoom buttons  /~~~~ */
+    if(zoomButtons){
+      let buttonContainer = chart.plotContainer.createChild(am4core.Container);
+      buttonContainer.shouldClone = false;
+      buttonContainer.align = zoomButtons.halign;
+      buttonContainer.valign = zoomButtons.valign;
+      buttonContainer.zIndex = Number.MAX_SAFE_INTEGER;
+      buttonContainer.marginTop = zoomButtons.marginTop;
+      if(zoomButtons.valign === "top"){
+        buttonContainer.marginTop = zoomButtons.marginV;
+      }else{
+        buttonContainer.marginBottom = zoomButtons.marginV;
+      }
+      if(zoomButtons.halign === "left"){
+        buttonContainer.marginLeft = zoomButtons.marginH;
+      }else{
+        buttonContainer.marginRight = zoomButtons.marginH;
+      }
+      buttonContainer.layout = "horizontal";
+      let zoomInButton = buttonContainer.createChild(am4core.Button);
+      zoomInButton.label.text = "+";
+      zoomInButton.events.on("hit", function(ev) {
+        let diff = XAxis.maxZoomed - XAxis.minZoomed;
+        let delta = diff * zoomButtons.zoomFactor;
+        if(isDate){
+          XAxis.zoomToDates(
+            new Date(XAxis.minZoomed + delta), new Date(XAxis.maxZoomed - delta)
+          );
+        }else{
+          XAxis.zoomToValues(
+            XAxis.minZoomed + delta, XAxis.maxZoomed - delta
+          );
+        }
+      });
+      let zoomOutButton = buttonContainer.createChild(am4core.Button);
+      zoomOutButton.label.text = "-";
+      zoomOutButton.events.on("hit", function(ev) {
+        let diff = XAxis.maxZoomed - XAxis.minZoomed;
+        let delta = diff * zoomButtons.zoomFactor;
+        if(isDate){
+          XAxis.zoomToDates(
+            new Date(XAxis.minZoomed - delta), new Date(XAxis.maxZoomed + delta)
+          );
+        }else{
+          XAxis.zoomToValues(
+            XAxis.minZoomed - delta, XAxis.maxZoomed + delta
+          );
+        }
+      });
+    }
+
+
+    /* ~~~~\                         /~~~~ */
     yValues.forEach(function (value, index) {
 
       let series = chart.series.push(new am4charts.LineSeries());